forth commit
I abandoned for now the idea of double-dispatch, I'll try to solve the problem without it and to come back later to this matter. I added print methond for output, finalized Pyromancer class, deleted and added new methods in Hero class, added game mechanism stuff in classes.

+++ src/com/oop_upb/modifiers/Modifiers.java
@@ -6,8 +6,8 @@ import com.oop_upb.heroes.Rogue;
 import com.oop_upb.heroes.Wizard;
 
 public interface Modifiers {
-    int modify(Knight knight);
-    int modify(Rogue rogue);
-    int modify(Wizard wizard);
-    int modify(Pyromancer pyromancer);
+    float modify(Knight knight);
+    float modify(Rogue rogue);
+    float modify(Wizard wizard);
+    float modify(Pyromancer pyromancer);
 }
